feat: add authentication components and refactor imports for better organization
