+++ blogSpace/src/components/functionality/CreatePost.jsx
